Split Engine specific logic from TestEnvironment
