+++ src/Tasks.UnitTests/Microsoft.Build.Tasks.UnitTests.csproj
@@ -51,6 +51,7 @@
     <Compile Include="..\Shared\UnitTests\ObjectModelHelpers.cs" />
     <Compile Include="..\Shared\UnitTests\ResourceUtilities_Tests.cs" />
     <Compile Include="..\Shared\QuotingUtilities.cs" />
+    <Compile Include="..\Shared\UnitTests\EngineTestEnvironment.cs" />
     <Compile Include="..\Shared\UnitTests\TestEnvironment.cs">
       <Link>TestEnvironment.cs</Link>
     </Compile>
