+++ src/Tasks.UnitTests/PortableTasks_Tests.cs
@@ -3,7 +3,7 @@
 
 using System.IO;
 using System.Text.RegularExpressions;
-using Microsoft.Build.Engine.UnitTests;
+using Microsoft.Build.UnitTests;
 using Microsoft.Build.Shared;
 using Microsoft.Build.UnitTests.Shared;
 using Shouldly;
