+++ src/Utilities.UnitTests/Microsoft.Build.Utilities.UnitTests.csproj
@@ -30,6 +30,7 @@
     <Compile Include="..\Shared\UnitTests\MockLogger.cs" />
     <Compile Include="..\Shared\UnitTests\ObjectModelHelpers.cs" />
     <Compile Include="..\Shared\UnitTests\ResourceUtilities_Tests.cs" />
+    <Compile Include="..\Shared\UnitTests\EngineTestEnvironment.cs" />
     <Compile Include="..\Shared\UnitTests\TestEnvironment.cs">
       <Link>TestEnvironment.cs</Link>
     </Compile>
