+++ src/Utilities/Microsoft.Build.Utilities.csproj
@@ -167,7 +167,7 @@
       <Link>VisualStudioLocationHelper.cs</Link>
     </Compile>
     <Compile Include="..\Shared\Traits.cs">
-      <Link>SharedTraits.cs</Link>
+      <Link>Traits.cs</Link>
     </Compile>
 
     <!-- Win32 RC Files -->
